use all double in eq_calc

+++ components/tas58xx/tas58xx_helpers.cpp
@@ -106,8 +106,8 @@ namespace esphome::tas58xx_helpers {
 
     double beta, b0, b1, b2, a1, a2;
 
-    double linear_gain = powf(10.0, static_cast<float>(gain) / 20.0);
-    double t0 = 2.0 * std::numbers::pi * frequency / static_cast<float>(sample_rate);
+    double linear_gain = powf(10.0, static_cast<double>(gain) / 20.0);
+    double t0 = 2.0 * std::numbers::pi * static_cast<double>frequency / static_cast<double>(sample_rate);
 
     if (linear_gain >= 1.0) {
         beta = t0 / (2.0 *  qFactor);
